gated-clocks: Now uses runmanager's load_shot_globals() function to get shot globals.
That function applies some hacks that we were otherwise applying in several
places. In particular a new hack allows us to use None as a global value.
It is stored as a null h5py.Reference object.

The function is backported to runmanager gtk so one need not upgrade runmanager.