GatedClocks-bugfix: Fixed bug in gated clocks feature
We previously only checked that instructions on clocklines were not too close together for the clock limit (i.e. was an instruction on clockline A too close to the next instruction on clockline A, and was instruction B too close to the next instruction on clockline B). However, this does not catch the case where an instruction on clockline B is close to an instruction on clockline A, thus forcing the pulse on clockline A to be shorter than it would otherwise.

This case is now caught in the below patch.